requirements.txt
python-dotenv>=0.16.0
requests~=2.25.1
dotmap~=1.3.23
rich==10.0.1
